echo: enable Werror
Part of #9460

* If there is an operation between signed and unsigned int, the unsigned is changed to signed one,
  except for some operations (if before the operation the signed one is checked to be less than
  zero or not, then the signed one get cast to be unsigned (errors due to sign error when casting
  are impossible))
* When doing operations on pointers, integers get cast to uintptr_t type

Untested: I don't have the hardware.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/7484
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/add-ons/kernel/drivers/audio/echo/multi.cpp
@@ -255,7 +255,7 @@ echo_get_mix(echo_dev *card, multi_mix_value_info * mmvi)
 	multi_mixer_control *control = NULL;
 	for (i = 0; i < mmvi->item_count; i++) {
 		id = mmvi->values[i].id - MULTI_CONTROL_FIRSTID;
-		if (id < 0 || id >= card->multi.control_count) {
+		if (id < 0 || (uint32)id >= card->multi.control_count) {
 			PRINT(("echo_get_mix : invalid control id requested : %" B_PRIu32
 				"\n", id));
 			continue;
@@ -297,7 +297,7 @@ echo_set_mix(echo_dev *card, multi_mix_value_info * mmvi)
 	multi_mixer_control *control = NULL;
 	for (i = 0; i < mmvi->item_count; i++) {
 		id = mmvi->values[i].id - MULTI_CONTROL_FIRSTID;
-		if (id < 0 || id >= card->multi.control_count) {
+		if (id < 0 || (uint32)id >= card->multi.control_count) {
 			PRINT(("echo_set_mix : invalid control id requested : %" B_PRIu32
 				"\n", id));
 			continue;
@@ -308,7 +308,7 @@ echo_set_mix(echo_dev *card, multi_mix_value_info * mmvi)
 			multi_mixer_control *control2 = NULL;
 			if (i + 1 < mmvi->item_count) {
 				id = mmvi->values[i + 1].id - MULTI_CONTROL_FIRSTID;
-				if (id < 0 || id >= card->multi.control_count) {
+				if (id < 0 || (uint32)id >= card->multi.control_count) {
 					PRINT(("echo_set_mix : invalid control id requested : %"
 						B_PRIu32 "\n", id));
 				} else {
